Windows: cm_user interlocked operations
Convert cm_user refcnts to use interlocked operations.

Reviewed-on: http://gerrit.openafs.org/9061
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Tested-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit e4d97fa3e4a0d02160e082547b84c437472bfaae)

Reviewed-on: http://gerrit.openafs.org/9063
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/WINNT/afsd/cm_user.h
@@ -44,10 +44,10 @@ typedef struct cm_ucell {
 #define CM_UCELLFLAG_ROOTUSER  0x10     /* cm_rootUser connection */
 
 typedef struct cm_user {
-    unsigned long refCount;             /* ref count - cm_userLock */
+    long refCount;                      /* ref count - cm_userLock */
     cm_ucell_t *cellInfop;	        /* list of cell info */
     osi_mutex_t mx;		        /* mutex */
-    int vcRefs;			        /* count of references from virtual circuits */
+    long vcRefs;			/* count of references from virtual circuits */
     long flags;
     GUID authgroup;                     /* AFS redirector */
 } cm_user_t;
